Crate libobs_wrapper

Modules

bootstrap

context

OBS Context Management

crash_handler

data

display

For this display method to work, another preview window has to be created in order to create a swapchain This is because the main window renderer is already handled by other processes

encoders

enums

logger

runtime

Runtime management for safe OBS API access across threads

scenes

signals

sources

unsafe_send

utils

Macros

impl_obs_drop

impl_signal_manager

mutex_blocking_lock

oneshot_rx_recv

run_with_obs

run_with_obs_blocking

This function can only be called OUTSIDE of the OBS thread.

run_with_obs_impl

rw_lock_blocking_read

rw_lock_blocking_write

rx_recv