libobs_wrapper::data::output

Struct ObsOutputRef

pub struct ObsOutputRef { /* private fields */ }

A reference to an OBS output.

This struct represents an output in OBS, which is responsible for outputting encoded audio and video data to a destination such as:

A file (recording)

A streaming service (RTMP, etc.)

A replay buffer

The output is associated with video and audio encoders that convert raw media to the required format before sending/storing.

Implementations

impl ObsOutputRef

pub fn settings(&self) -> &Arc<RwLock<Option<ObsData>>>

pub fn hotkey_data(&self) -> &Arc<RwLock<Option<ObsData>>>

pub fn video_encoders(&self) -> &Arc<RwLock<Vec<Arc<ObsVideoEncoder>>>>

pub fn audio_encoders(&self) -> &Arc<RwLock<Vec<Arc<ObsAudioEncoder>>>>

pub fn id(&self) -> &ObsString

pub fn name(&self) -> &ObsString

pub fn signal_manager(&self) -> &Arc<ObsOutputSignals>

pub fn video_encoders_mut( &mut self, ) -> &mut Arc<RwLock<Vec<Arc<ObsVideoEncoder>>>>

pub fn audio_encoders_mut( &mut self, ) -> &mut Arc<RwLock<Vec<Arc<ObsAudioEncoder>>>>

impl ObsOutputRef

pub async fn get_video_encoders(&self) -> Vec<Arc<ObsVideoEncoder>>

Returns a list of all video encoders attached to this output.

Returns

A vector of Arc-wrapped ObsVideoEncoder instances

pub async fn video_encoder( &mut self, info: VideoEncoderInfo, handler: Sendable<*mut video_output>, ) -> Result<Arc<ObsVideoEncoder>, ObsError>

Creates and attaches a new video encoder to this output.

This method creates a new video encoder using the provided information, sets up the video handler, and attaches it to this output.

Arguments

info - Information for creating the video encoder

handler - The video output handler

Returns

A Result containing an Arc-wrapped ObsVideoEncoder or an error

pub async fn set_video_encoder( &mut self, encoder: ObsVideoEncoder, ) -> Result<(), ObsError>

Attaches an existing video encoder to this output.

Arguments

encoder - The video encoder to attach

Returns

A Result indicating success or an error

pub async fn update_settings( &mut self, settings: ObsData, ) -> Result<(), ObsError>

Updates the settings of this output.

Note: This can only be done when the output is not active.

Arguments

settings - The new settings to apply

Returns

A Result indicating success or an error

pub async fn audio_encoder( &mut self, info: AudioEncoderInfo, mixer_idx: usize, handler: Sendable<*mut audio_output>, ) -> Result<Arc<ObsAudioEncoder>, ObsError>

Creates and attaches a new audio encoder to this output.

This method creates a new audio encoder using the provided information, sets up the audio handler, and attaches it to this output at the specified mixer index.

Arguments

info - Information for creating the audio encoder

mixer_idx - The mixer index to use (typically 0 for primary audio)

handler - The audio output handler

Returns

A Result containing an Arc-wrapped ObsAudioEncoder or an error

pub async fn set_audio_encoder( &mut self, encoder: ObsAudioEncoder, mixer_idx: usize, ) -> Result<(), ObsError>

Attaches an existing audio encoder to this output at the specified mixer index.

Arguments

encoder - The audio encoder to attach

mixer_idx - The mixer index to use (typically 0 for primary audio)

Returns

A Result indicating success or an error

pub async fn start(&self) -> Result<(), ObsError>

Starts the output.

This begins the encoding and streaming/recording process.

Returns

A Result indicating success or an error (e.g., if the output is already active)

pub async fn stop(&mut self) -> Result<(), ObsError>

Stops the output.

This ends the encoding and streaming/recording process. The method waits for a stop signal and returns the result.

Returns

A Result indicating success or an error with details about why stopping failed

pub fn as_ptr(&self) -> Sendable<*mut obs_output>