libobs_wrapper

Module display

For this display method to work, another preview window has to be created in order to create a swapchain This is because the main window renderer is already handled by other processes

Structs

DisplayWindowManager

ObsDisplayCreationData

ObsDisplayRef

NEVER STORE THIS REF DIRECTLY!!

Enums

GsColorFormat

GsZstencilFormat

Traits

MiscDisplayTrait

ShowHideTrait

WindowPositionTrait