libobs_wrapper

Module enums

Enums

ObsColorspace

Describes the colorspace that an OBS video context uses. Used in ObsVideoInfo.

ObsEncoderType

ObsGraphicsModule

Describes which graphics backend should be used in the OBS video context. Used in ObsVideoInfo.

ObsLogLevel

ObsOutputStopSignal

ObsResetVideoStatus

Status types returned after attempting to reset the OBS video context using the function obs_reset_video.

ObsSamplesPerSecond

Audio samples per second options that are supported by libobs.

ObsScaleType

Describes how libobs should reconcile non-matching base and output resolutions when creating a video context.

ObsSpeakerLayout

ObsVideoFormat

Describes the video output format used by the OBS video context. Used in ObsVideoInfo.

ObsVideoRange

Describes the minimum and maximum color levels that an OBS video context is allowed to encode. Used in ObsVideoInfo.